pyproject.toml
[tool.poetry]
name = "shuffle"
version = "0.1.0"
description = ""
authors = ["Tiago Requeijo <tiago.requeijo.dev@gmail.com>"]
license = "MIT"
build = "build.py"
#include = ["*.hpp"]

[tool.poetry.dependencies]
python = "^3.4"
numpy = "^1.16"
cython = "^0.29.3"
python-snappy = {version = "^0.5.3",optional = true}

[tool.poetry.dev-dependencies]
pytest-cov = "^2.6"
pytest = "^4.1"
pytest-flake8 = "^1.0"
pytest-mypy = "^0.3.2"

[tool.poetry.extras]
snappy = ["python-snappy"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
